Refactore edge pool related functions into separate files
